[MNG-6511] Extract ActivationSettings as separate enum
